FEATURE: Initial work on changes to resource unpacking
