+++ Forge/src/main/java/io/github/lukebemish/defaultresources/impl/forge/DefaultResourcesForge.java
@@ -1,10 +1,10 @@
-package io.github.lukebemish.defaultresources.forge;
+package io.github.lukebemish.defaultresources.impl.forge;
 
 import com.mojang.datafixers.util.Pair;
-import io.github.lukebemish.defaultresources.AutoMetadataFilePackResources;
-import io.github.lukebemish.defaultresources.AutoMetadataFolderPackResources;
-import io.github.lukebemish.defaultresources.DefaultResources;
-import io.github.lukebemish.defaultresources.Services;
+import io.github.lukebemish.defaultresources.impl.AutoMetadataFilePackResources;
+import io.github.lukebemish.defaultresources.impl.AutoMetadataFolderPackResources;
+import io.github.lukebemish.defaultresources.impl.DefaultResources;
+import io.github.lukebemish.defaultresources.impl.Services;
 import io.github.lukebemish.defaultresources.api.ResourceProvider;
 import net.minecraft.SharedConstants;
 import net.minecraft.network.chat.Component;
